src/telran/util/Words.java
package telran.util;
import java.util.*;

public class Words {
	TreeSet<String> set = new TreeSet<>(String.CASE_INSENSITIVE_ORDER);
	/**
	 * adds word
	 * @param word
	 * @return true if added, otherwise false if an word already exists (case insensitive)
	 */
	public boolean addWord(String word) {		
		return set.add(word);
	}
	/**
	 * 
	 * @param prefix
	 * @return list of words starting from a given prefix (case insensitive)
	 */
	public List<String> getWordsByPrefix(String prefix) {		
		return new ArrayList<>(set.subSet(prefix, getPrefixLimit(prefix)));
	}
	
	private String getPrefixLimit(String prefix) {
		char lastChar = prefix.charAt(prefix.length() - 1);
		char newLastChar = (char) (lastChar + 1);
		return prefix.substring(0, prefix.length() - 1) + newLastChar;
	}
}